+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-29T14:44:57.871Z")
+    created: Optional[datetime] = parse_datetime("2025-09-29T16:10:26.831Z")
     r"""The date and time the resource was created"""