+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-29T14:44:54.875Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-29T16:10:23.816Z")
     r"""The date and time the resource was last updated"""