+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6AYYBBRF3Q1VC4GMRR14MCG"
+        "01K6B3TW06CQ8YR498JFMRH7BF"
     )
     r"""The id of the resource"""