+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -571,7 +571,7 @@ class HTTPTool(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6AYYB85CAWZ294ZD55V04D3"
+        "01K6B3TVWQRGGQ37MWVXXVDPCP"
     )
 
     requires_approval: Optional[bool] = False